+++ src/RuleInterface.php
@@ -15,9 +15,9 @@ interface RuleInterface
      * @param string|null $userId The user ID. This should be a string representing the unique identifier of a user. For
      * guests the value is `null`.
      * @param Item $item The role or permission that this rule is associated with.
-     * @param array $parameters Parameters passed to {@see CheckAccessInterface::userHasPermission()}.
+     * @param RuleContext $ruleContext Rule context.
      *
      * @return bool Whether the rule permits the auth item it is associated with.
      */
-    public function execute(?string $userId, Item $item, array $parameters = []): bool;
+    public function execute(?string $userId, Item $item, RuleContext $ruleContext): bool;
 }