+++ tests/Support/EasyRule.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Rbac\Tests\Support;
 
 use Yiisoft\Rbac\Item;
+use Yiisoft\Rbac\RuleContext;
 use Yiisoft\Rbac\RuleInterface;
 
 final class EasyRule implements RuleInterface
@@ -13,7 +14,7 @@ final class EasyRule implements RuleInterface
     {
     }
 
-    public function execute(?string $userId, Item $item, array $parameters = []): bool
+    public function execute(?string $userId, Item $item, RuleContext $ruleContext): bool
     {
         return $this->expected;
     }