+++ CHANGELOG.md
@@ -28,7 +28,8 @@
     (@arogachev)
 - Enh #165: Rename `DefaultRoleNotFoundException` to `DefaultRolesNotFoundException` and finalize it (@arogachev)
 - Bug #172: Execute rule when checking permissions for guests (@arogachev)
-- Chg #172: Make `$userId` parameter `nullable` in `RuleInterface::execute()` (@arogachev) 
+- Chg #172: Make `$userId` parameter `nullable` in `RuleInterface::execute()` (@arogachev)
+- Bug #175: Use rule factory for creating rule instances in `CompositeRule` (@arogachev)
 
 ## 1.0.2 April 20, 2023